Renamed movie card potrait
