--- shared/src/commonMain/kotlin/com/vickikbt/shared/ui/components/ItemSimilarMovies.kt
@@ -1,39 +1,0 @@
-package com.vickikbt.shared.ui.components
-
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.unit.dp
-import com.seiko.imageloader.rememberAsyncImagePainter
-import com.vickikbt.shared.domain.models.Movie
-import com.vickikbt.shared.ui.components.ratingbar.RatingBar
-import com.vickikbt.shared.ui.components.ratingbar.RatingBarStyle
-import com.vickikbt.shared.ui.components.ratingbar.StepSize
-import com.vickikbt.shared.utils.getRating
-import com.vickikbt.shared.utils.loadImage
-
-@Composable
-fun ItemSimilarMovies(movie: Movie, onItemClick: (Movie) -> Unit) {
-    val painter = rememberAsyncImagePainter(movie.posterPath?.loadImage() ?: "")
-
-    Column(verticalArrangement = Arrangement.spacedBy(1.dp)) {
-
-        MovieCardPortrait(
-            //modifier = Modifier.height(220.dp),
-            movie = movie,
-            onItemClick = { onItemClick(movie) })
-
-        // rating bar
-        RatingBar(
-            modifier = Modifier.fillMaxWidth(0.8f),
-            value = movie.voteAverage?.getRating()?.toFloat() ?: 0f,
-            numOfStars = 5,
-            size = 15.dp,
-            stepSize = StepSize.HALF,
-            isIndicator = true,
-            style = RatingBarStyle.Fill()
-        )
-    }
-}
